feat: Enhance chat logs and metrics handling
- Updated chatLogsHandler to support pagination with lastId and limit parameters.
- Added totalCount to the response for chat logs.
- Modified ChatLogsDashboard to fetch logs in chunks and display loading progress.
- Refactored MetricsDashboard to use MetricsService for fetching chat logs and calculating metrics.
- Improved error handling and loading indicators in both dashboards.
- Removed deprecated DataStoreService methods related to chat logs.
- Added new translations for loading indicators in English and French.
