+++ src/locales/en.json
@@ -349,7 +349,8 @@
       "interactionsFound": "chats. Download the logs to see the full set and details.",
       "date": "Date",
       "chatId": "Chat ID",
-      "interactions": "Interactions"
+      "interactions": "Interactions",
+      "loadingRecordsLeft": "Loading logs: {recordsLeft} records left to load ({totalCount} total records)"
     },
     "deleteChat": {
       "title": "Delete a chat with incorrect feedback",
